refactor(map): extract view-model builder into lib/src/view/ (#3574)
Continues slice 5 of the colonizethis_map wave-2 refactor: after the
render layer moved to lib/src/render/, relocate the view-model building
cluster into lib/src/view/, establishing the view layer of the explicit
generation -> view inputs -> render layering.

Moved (import-path-only, no logic change):
- init_game_map_view_builder.dart (+ fleet/orchestration/cells/map-markers
  part files)
- init_game_map_view_data.dart
- region_map_view_inputs.dart

Fixed sibling/relative imports in the moved files (../ for flat src/
modules, ../render/ for render siblings, same-dir for the co-moved view
files) and updated their importers (render modules, gp/province tints,
the public colonizethis_map.dart barrel re-exports, and the same-package
region_map_view_inputs test). Consumers (app/, ctdev/, tool/,
colonizethis_setup) import via the barrel and are unaffected.

SPEC: document the view layer alongside the render layer in
SPEC/program/map-visualization.md (Responsibility), noting the
gen -> view -> render dependency direction.

Verification: dart analyze (only 4 pre-existing unrelated test warnings),
dart test (296 passing incl. determinism, locked pair, visualization
goldens, region_map_view_inputs), and all eight repo.map_* lint checks
(public_barrel_surface, gen_no_image_import, no_partfile_classes,
region_dispatch_central, grid_cell_iteration_central,
region_data_access_central, grid_ops_central, gen_stage_protocol) green.

Refs #3574
